feat: moved summary information to common component
