+++ src/Components/ActionsBuilder/CallDetails/CallDetails.styled.tsx
@@ -1,5 +1,5 @@
 import { MetadataTag } from '../SupportedActions/common/infoLine';
-import { DetailRow } from '../SupportedActions/common/summary';
+import { DetailRow } from '../SupportedActions/common/Summary.styled';
 import styled, { css } from 'styled-components';
 import { Box } from 'Components/Primitives/Layout';
 import { Button } from 'old-components/Guilds/common/Button';
