+++ src/Components/ActionsBuilder/SupportedActions/ERC20Transfer/ERC20TransferSummary.tsx
@@ -1,15 +1,10 @@
 import { ActionViewProps } from '..';
-import { Segment } from '../common/infoLine';
-import { DetailBody, DetailHeader, DetailRow } from '../common/summary';
 import { BigNumber } from 'ethers';
-import useENSAvatar from 'hooks/Guilds/ether-swr/ens/useENSAvatar';
-import Avatar from 'old-components/Guilds/Avatar';
 import { useMemo } from 'react';
-import { MAINNET_ID } from 'utils';
-import { useTranslation } from 'react-i18next';
+import Summary from '../common/Summary';
 
 const ERC20TransferSummary: React.FC<ActionViewProps> = ({ decodedCall }) => {
-  const { t } = useTranslation();
+  console.log(decodedCall);
 
   const parsedData = useMemo(() => {
     if (!decodedCall) return null;
@@ -22,27 +17,9 @@ const ERC20TransferSummary: React.FC<ActionViewProps> = ({ decodedCall }) => {
     };
   }, [decodedCall]);
 
-  const { ensName, imageUrl } = useENSAvatar(
-    parsedData?.destination,
-    MAINNET_ID
-  );
-
   return (
     <>
-      <DetailHeader>{t('interactWith')}:</DetailHeader>
-
-      <DetailRow>
-        <DetailBody>
-          <Segment>
-            <Avatar
-              defaultSeed={parsedData?.destination}
-              src={imageUrl}
-              size={24}
-            />
-          </Segment>
-          <Segment>{ensName || parsedData?.destination}</Segment>
-        </DetailBody>
-      </DetailRow>
+      <Summary decodedCall={decodedCall} address={parsedData?.destination} />
     </>
   );
 };
